requirements.txt
pandas
prophet
flask
flask-cors
numpy
python-dotenv
pocketbase